image and file upload

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
@@ -58,8 +59,27 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        # handle file upload
+        if ($request->hasFile('cover_image')) {
+            if ($request->file('cover_image')->isValid()) {
+                # get file name with extension
+                # $fielNameWithExt = $request->file('cover_image');
+                # get just filename
+                # $fileName = pathinfo($fielNameWithExt, PATHINFO_FILENAME);
+                # get just ext
+                # $extension = $request->file('cover_image')->extension();
+                # file name to store
+                $fileNameToStore = $request->file('cover_image')->hashName();
+                # upload image
+                $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            }
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         # Create post
 
@@ -67,6 +87,7 @@ class PostsController extends Controller
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
 
         if ($post->save()) {
             return redirect('/posts')->with('success', 'Post Created');
@@ -115,14 +136,34 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        # handle file upload
+        if ($request->hasFile('cover_image')) {
+            if ($request->file('cover_image')->isValid()) {
+                # get file name with extension
+                # $fielNameWithExt = $request->file('cover_image');
+                # get just filename
+                # $fileName = pathinfo($fielNameWithExt, PATHINFO_FILENAME);
+                # get just ext
+                # $extension = $request->file('cover_image')->extension();
+                # file name to store
+                $fileNameToStore = $request->file('cover_image')->hashName();
+                # upload image
+                $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            }
+        }
 
         # Update post
 
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
 
         if ($post->save()) {
             return redirect('/posts')->with('success', 'Post Updated');
@@ -144,11 +185,16 @@ class PostsController extends Controller
         if (auth()->user()->id !== $post->user_id) {
             return redirect('/posts')->with('error', 'Unauthorized Page');
         } else {
-            if ($post->delete()) {
-                return redirect('/posts')->with('success', 'Post Deleted');
-            } else {
-                return view('/posts')->with('error', 'There was an error deleting your post');
+            if ($post->cover_image != 'noimage.jpg') {
+                Storage::delete('public/cover_images/'.$post->cover_image);
+
+                if ($post->delete()) {
+                    return redirect('/posts')->with('success', 'Post Deleted');
+                } else {
+                    return view('/posts')->with('error', 'There was an error deleting your post');
+                }
             }
+            
         }
